Refactored Stopwatch::SetFrequency and Stopwatch::ctor for frequency to use unsigned int;

+++ Engine/Code/Engine/Core/Console.hpp
@@ -123,9 +123,9 @@ private:
     std::string::const_iterator _cursor_position{};
     std::string::const_iterator _selection_position{};
     decltype(_entryline_buffer)::const_iterator _current_history_position{};
-    int _default_blink_rate = 3;
-    int _blink_rate = _default_blink_rate;
-    Stopwatch _cursor_timer = Stopwatch(static_cast<float>(_blink_rate));
+    unsigned int _default_blink_rate = 3u;
+    unsigned int _blink_rate = _default_blink_rate;
+    Stopwatch _cursor_timer = Stopwatch(_blink_rate);
     uint8_t _show_cursor             : 1;
     uint8_t _is_open                 : 1;
     uint8_t _highlight_mode          : 1;
